back/src/requirements.txt
django-pwa==1.0.5
django-webpush==0.2.7
py-vapid==1.5.0
django-bootstrap-datepicker-plus==3.0.5